Add closure to build products
The idea is that the "install" attribute is what gets installed in the
packet-broker profile on the installation target. It includes the
services and the wrapper for the local kernel.  The "closure"
attribute is what needs to be copied to a cache to implement a binary
deployment.  It includes the modules for all kernels but doesn't
include neither the wrapper nor the service unit files.